+++ library/src/schemas/literal/literal.ts
@@ -1,16 +1,16 @@
 import type { BaseSchema } from '../../types.ts';
 import { getIssue } from '../../utils/index.ts';
-import type { Primitive } from './types.ts';
+import type { Literal } from './types.ts';
 
 /**
  * Literal schema type.
  */
 export type LiteralSchema<
-  TLiteralValue extends Primitive,
-  TOutput = TLiteralValue
-> = BaseSchema<TLiteralValue, TOutput> & {
+  TLiteral extends Literal,
+  TOutput = TLiteral
+> = BaseSchema<TLiteral, TOutput> & {
   schema: 'literal';
-  literal: TLiteralValue;
+  literal: TLiteral;
 };
 
 /**
@@ -21,7 +21,7 @@ export type LiteralSchema<
  *
  * @returns A literal schema.
  */
-export function literal<TLiteral extends Primitive>(
+export function literal<TLiteral extends Literal>(
   literal: TLiteral,
   error?: string
 ): LiteralSchema<TLiteral> {