+++ library/src/schemas/literal/literalAsync.ts
@@ -1,16 +1,16 @@
 import type { BaseSchemaAsync } from '../../types.ts';
 import { getIssue } from '../../utils/index.ts';
-import type { Primitive } from './types.ts';
+import type { Literal } from './types.ts';
 
 /**
  * Literal schema async type.
  */
 export type LiteralSchemaAsync<
-  TLiteralValue extends Primitive,
-  TOutput = TLiteralValue
-> = BaseSchemaAsync<TLiteralValue, TOutput> & {
+  TLiteral extends Literal,
+  TOutput = TLiteral
+> = BaseSchemaAsync<TLiteral, TOutput> & {
   schema: 'literal';
-  literal: TLiteralValue;
+  literal: TLiteral;
 };
 
 /**
@@ -21,7 +21,7 @@ export type LiteralSchemaAsync<
  *
  * @returns An async literal schema.
  */
-export function literalAsync<TLiteral extends Primitive>(
+export function literalAsync<TLiteral extends Literal>(
   literal: TLiteral,
   error?: string
 ): LiteralSchemaAsync<TLiteral> {